Implement secp256k1 GPG key generation
