+++ tests/test_bip85_gpg.py
@@ -27,6 +27,6 @@ def test_bip85_secp256k1_deterministic():
     seed = Seed(mnemonic=MNEMONIC)
     root = bip32.HDKey.from_seed(seed.seed_bytes)
     key = bip85_secp256k1_from_root(root, 0)
-    assert int(key.s) == int(
+    assert int.from_bytes(key.secret, "big") == int(
         "cefbb3197f44cbcd28ca548e7d6c22e2b67f497caeebb71fa91d1cc6ab78e502", 16
     )
